Fix the single file libraries being copied as source.
The code generates a temp file with the library version.
By mistake that was ALWAYS copied back to the mpy file for single file libraries.
The temporary files for the package libraries were also not deleted.

For windows compatibility, the temporary file must be copied (and therefore erased) outside of the with block. In this PR we only keep the temporary file open the time to write to it, and then do the rest.